Refactoring: move some classes in more suitable packages
