+++ gridcapa-task-manager-app/src/main/java/com/farao_community/farao/gridcapa/task_manager/app/service/ParameterService.java
@@ -8,7 +8,7 @@ package com.farao_community.farao.gridcapa.task_manager.app.service;
 
 import com.farao_community.farao.gridcapa.task_manager.api.ParameterDto;
 import com.farao_community.farao.gridcapa.task_manager.api.TaskManagerException;
-import com.farao_community.farao.gridcapa.task_manager.app.ParameterRepository;
+import com.farao_community.farao.gridcapa.task_manager.app.repository.ParameterRepository;
 import com.farao_community.farao.gridcapa.task_manager.app.configuration.RunnerParameters;
 import com.farao_community.farao.gridcapa.task_manager.app.entities.Parameter;
 import org.slf4j.Logger;
